refactor: move system tests into `tests/system` directory
This aligns more closely with the templates from synthtool.
