+++ tests/system/test_client.py
@@ -25,10 +25,8 @@ import os
 import time
 import unittest
 import uuid
-import re
 
 import psutil
-import pytest
 import pytz
 import pkg_resources
 
@@ -51,13 +49,6 @@ try:
     import pyarrow.types
 except ImportError:  # pragma: NO COVER
     pyarrow = None
-try:
-    import IPython
-    from IPython.utils import io as ipython_io
-    from IPython.testing import tools
-    from IPython.terminal import interactiveshell
-except ImportError:  # pragma: NO COVER
-    IPython = None
 
 from google.api_core.exceptions import PreconditionFailed
 from google.api_core.exceptions import BadRequest
@@ -2979,47 +2970,6 @@ class TestBigQuery(unittest.TestCase):
         return dataset
 
 
-@pytest.mark.skipif(pandas is None, reason="Requires `pandas`")
-@pytest.mark.skipif(IPython is None, reason="Requires `ipython`")
-@pytest.mark.usefixtures("ipython_interactive")
-def test_bigquery_magic():
-    ip = IPython.get_ipython()
-    current_process = psutil.Process()
-    conn_count_start = len(current_process.connections())
-
-    ip.extension_manager.load_extension("google.cloud.bigquery")
-    sql = """
-        SELECT
-            CONCAT(
-            'https://stackoverflow.com/questions/',
-            CAST(id as STRING)) as url,
-            view_count
-        FROM `bigquery-public-data.stackoverflow.posts_questions`
-        WHERE tags like '%google-bigquery%'
-        ORDER BY view_count DESC
-        LIMIT 10
-    """
-    with ipython_io.capture_output() as captured:
-        result = ip.run_cell_magic("bigquery", "--use_rest_api", sql)
-
-    conn_count_end = len(current_process.connections())
-
-    lines = re.split("\n|\r", captured.stdout)
-    # Removes blanks & terminal code (result of display clearing)
-    updates = list(filter(lambda x: bool(x) and x != "\x1b[2K", lines))
-    assert re.match("Executing query with job ID: .*", updates[0])
-    assert all(re.match("Query executing: .*s", line) for line in updates[1:-1])
-    assert re.match("Query complete after .*s", updates[-1])
-    assert isinstance(result, pandas.DataFrame)
-    assert len(result) == 10  # verify row count
-    assert list(result) == ["url", "view_count"]  # verify column names
-
-    # NOTE: For some reason, the number of open sockets is sometimes one *less*
-    # than expected when running system tests on Kokoro, thus using the <= assertion.
-    # That's still fine, however, since the sockets are apparently not leaked.
-    assert conn_count_end <= conn_count_start  # system resources are released
-
-
 def _job_done(instance):
     return instance.state.lower() == "done"
 
@@ -3039,21 +2989,3 @@ def _table_exists(t):
         return True
     except NotFound:
         return False
-
-
-@pytest.fixture(scope="session")
-def ipython():
-    config = tools.default_config()
-    config.TerminalInteractiveShell.simple_prompt = True
-    shell = interactiveshell.TerminalInteractiveShell.instance(config=config)
-    return shell
-
-
-@pytest.fixture()
-def ipython_interactive(request, ipython):
-    """Activate IPython's builtin hooks
-
-    for the duration of the test scope.
-    """
-    with ipython.builtin_trap:
-        yield ipython
